*Artist parameters (shader constants, etc.) can now be modified at run-time by using the arrow buttons *Rendering will now pause when application is not in focus *Added high resolution timer for better delta time precision *Added support for multithreaded resource loading *32-bit binaries are now exported in the 'Bin/x86/' folder (instead of Win32) *Moved all platform specific code to Framework class *Added Gainput library for platform independent input handling *Added FreeType 2.6 library for rendering on-screen text *Created a new pass which uses FreeType 2.6 to render text in a texture which then gets drawn on-screen *Removed the depth bias from DeferrredLightDir shader in favor of depth bias and slope scale depth bias render states set before rendering each cascade *Fixed a bug in RSMUpscale shader *Added a NoPCF() utility function in Utils.hlsl which has the same prototype as the other PCF functions *Fixed some shadow map issues *Shaders are now compiled during the build process via a custom build step in order for errors to be easier to fix; the outputs are assembly vertex and pixel programs (*vsasm *psasm) which are not used by the actual program, but allow Visual Studio to decide whether or not to recompile shaders (it's a lengthy process) *Disabled bokeh DoF effect by default since it is very expensive *Added a font file (Arial) *Made some minor modifications to project files with regard to intermediary and output directories *Removed debug information from Release builds *Added some checks for null pointers in ResourceManager class *Fixed some invalid file references in LibRenderer project (it did not affect the build process, but it would cause it to always rebuild) *Organized file filters in all projects

+++ GITechDemo/Code/Libraries/LibRenderer/Utility/HalfFloat.cpp
@@ -34,9 +34,8 @@ HalfFloat::HalfFloat(const float value)
 			if (biasExp != 0)
 			{
 				// Truncate 23-bit mantissa to 10 bits.
-				unsigned short signBit =
-					(unsigned short)((bits & 0x80000000) >> 16);
-				unsigned int mantissa = (bits & 0x007FFFFF) >> 13;
+				unsigned short signBit	= (unsigned short)((bits & 0x80000000) >> 16);
+				unsigned short mantissa	= (unsigned short)((bits & 0x007FFFFF) >> 13);
 				biasExp = (biasExp - 0x0070) << 10;
 				m_hValue = signBit | biasExp | mantissa;
 			}
@@ -81,6 +80,3 @@ HalfFloat::operator float() const
 		return 0.0f;
 	}
 }
-
-
-
